+++ tests/testthat/test-intervals.R
@@ -25,4 +25,11 @@ test_that("Check maximum interval", {
   expect_equal(max(TEST_INTER), 30)
 })
 
+test_that("quantile.intervals() basic cases", {
+  expect_equal(quantile(TEST_INTER), 
+               c(3.918526, 10.590013, 20.555245), 
+               tolerance=1e-3)
+  expect_equal(quantile(TEST_INTER, q = c(0.25)), 
+               c(6.227954), tolerance=1e-3)
+})